tester: simplify
