+++ Cargo.toml
@@ -24,7 +24,7 @@ members = [
     "kinode/packages/terminal/terminal",
     "kinode/packages/terminal/alias", "kinode/packages/terminal/cat", "kinode/packages/terminal/echo", "kinode/packages/terminal/hi", "kinode/packages/terminal/m", "kinode/packages/terminal/top",
     "kinode/packages/terminal/namehash_to_name", "kinode/packages/terminal/net_diagnostics", "kinode/packages/terminal/peer", "kinode/packages/terminal/peers",
-    "kinode/packages/tester/tester", "kinode/packages/tester/test_runner",
+    "kinode/packages/tester/tester",
 ]
 default-members = ["lib"]
 resolver = "2"
