+++ kinode/packages/tester/api/tester:sys-v0.wit
@@ -1,0 +1,27 @@
+interface tester {
+    variant request {
+        run(run-request),
+    }
+
+    variant response {
+        run(result<_, fail-response>)
+    }
+
+    record run-request {
+        input-node-names: list<string>,
+        test-names: list<string>,
+        test-timeout: u64,
+    }
+
+    record fail-response {
+        test: string,
+        file: string,
+        line: u32,
+        column: u32,
+    }
+}
+
+world tester-sys-v0 {
+    import tester;
+    include process-v0;
+}
