--- kinode/packages/tester/test_runner/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "test_runner"
-version = "0.1.0"
-edition = "2021"
-
-[features]
-simulation-mode = []
-
-[dependencies]
-anyhow = "1.0"
-bincode = "1.3.3"
-kinode_process_lib = { git = "https://github.com/kinode-dao/process_lib", rev = "09dc9f9" }
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-thiserror = "1.0"
-wit-bindgen = "0.24.0"
-
-[lib]
-crate-type = ["cdylib"]
-
-[package.metadata.component]
-package = "kinode:process"
