+++ kinode/packages/tester/tester/Cargo.toml
@@ -9,8 +9,8 @@ simulation-mode = []
 [dependencies]
 anyhow = "1.0"
 bincode = "1.3.3"
-indexmap = "2.1"
 kinode_process_lib = { git = "https://github.com/kinode-dao/process_lib", rev = "09dc9f9" }
+process_macros = { git = "https://github.com/kinode-dao/process_macros", rev = "626e501" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 thiserror = "1.0"
