+++ kinode/packages/tester/tester/src/lib.rs
@@ -1,20 +1,23 @@
-use indexmap::map::IndexMap;
+use std::collections::HashMap;
+use std::str::FromStr;
 
+use crate::kinode::process::tester::{
+    Request as TesterRequest, Response as TesterResponse, RunRequest, FailResponse,
+};
 use kinode_process_lib::kernel_types as kt;
 use kinode_process_lib::{
     await_message, call_init, our_capabilities, println, spawn, vfs, Address, Message, OnExit,
     ProcessId, Request, Response,
 };
 
-mod tester_types;
-use tester_types as tt;
+mod tester_lib;
 
 wit_bindgen::generate!({
     path: "target/wit",
-    world: "process-v0",
+    world: "tester-sys-v0",
+    generate_unused_types: true,
+    additional_derives: [PartialEq, serde::Deserialize, serde::Serialize, process_macros::SerdeJsonInto],
 });
-
-type Messages = IndexMap<kt::Message, tt::KernelMessage>;
 
 fn make_vfs_address(our: &Address) -> anyhow::Result<Address> {
     Ok(Address {
@@ -23,88 +26,174 @@ fn make_vfs_address(our: &Address) -> anyhow::Result<Address> {
     })
 }
 
+fn handle_response(message: &Message) -> anyhow::Result<()> {
+    let TesterResponse::Run(_) = message.body().try_into()?;
+    let source = message.source();
+    if (source.process.package_name != "tester") || (source.process.publisher_node != "sys") {
+        println!(
+            "got Response from unexpected source: {}; must be in package test:sys",
+            source,
+        );
+        fail!("tester");
+    }
+    Response::new().body(message.body()).send().unwrap();
+    Ok(())
+}
+
+fn read_caps_by_child(
+    dir_prefix: &str,
+    children: &mut Vec<vfs::DirEntry>,
+) -> anyhow::Result<HashMap<String, Vec<String>>> {
+    let caps_file_path = format!("{}/grant_capabilities.json", dir_prefix);
+    let caps_index = children.iter().position(|i| *i.path == *caps_file_path);
+    let caps_by_child: HashMap<String, Vec<String>> =
+        match caps_index {
+            None => HashMap::new(),
+            Some(caps_index) => {
+                children.remove(caps_index);
+                let file = vfs::file::open_file(&caps_file_path, false, None)?;
+                let file_contents = file.read()?;
+                serde_json::from_slice(&file_contents)?
+            }
+        };
+    Ok(caps_by_child)
+}
+
+fn handle_request(
+    our: &Address,
+    message: &Message,
+    node_names: &mut Vec<String>,
+) -> anyhow::Result<()> {
+    let TesterRequest::Run(RunRequest {
+        input_node_names,
+        ref test_names,
+        test_timeout,
+    }) = message.body().try_into()?;
+    println!("got Run");
+
+    assert!(input_node_names.len() >= 1);
+    *node_names = input_node_names.clone();
+
+    if our.node != node_names[0] {
+        // we are not the master node
+        Response::new()
+            .body(TesterResponse::Run(Ok(())))
+            .send()
+            .unwrap();
+        return Ok(());
+    }
+
+    // we are the master node
+    let dir_prefix = "tester:sys/tests";
+
+    let response = Request::new()
+        .target(make_vfs_address(&our)?)
+        .body(serde_json::to_vec(&vfs::VfsRequest {
+            path: dir_prefix.into(),
+            action: vfs::VfsAction::ReadDir,
+        })?)
+        .send_and_await_response(test_timeout)??;
+
+    let Message::Response { body: vfs_body, .. } = response else {
+        fail!("tester");
+    };
+    let vfs::VfsResponse::ReadDir(mut children) =
+        serde_json::from_slice(&vfs_body)?
+    else {
+        println!(
+            "{:?}",
+            serde_json::from_slice::<serde_json::Value>(&vfs_body)?
+        );
+        fail!("tester");
+    };
+
+    for test_name in test_names {
+        let test_entry = vfs::DirEntry {
+            path: format!("{}/{}.wasm", dir_prefix, test_name),
+            file_type: vfs::FileType::File,
+        };
+        if !children.contains(&test_entry) {
+            return Err(anyhow::anyhow!(
+                "test {} not found amongst {:?}",
+                test_name,
+                children,
+            ));
+        }
+    }
+
+    let caps_by_child = read_caps_by_child(dir_prefix, &mut children)?;
+
+    println!("tester: running {:?}...", children);
+
+    for test_name in test_names {
+        let test_path = format!("{}/{}.wasm", dir_prefix, test_name);
+        let grant_caps = caps_by_child
+            .get(test_name)
+            .and_then(|caps| {
+                Some(
+                    caps.iter()
+                        .map(|cap| ProcessId::from_str(cap).unwrap())
+                        .collect(),
+                )
+            })
+            .unwrap_or(vec![]);
+        let child_process_id = match spawn(
+            None,
+            &test_path,
+            OnExit::None, //  TODO: notify us
+            our_capabilities(),
+            grant_caps,
+            false, // not public
+        ) {
+            Ok(child_process_id) => child_process_id,
+            Err(e) => {
+                println!("couldn't spawn {}: {}", test_path, e);
+                fail!("tester");
+            }
+        };
+
+        let response = Request::new()
+            .target(Address {
+                node: our.node.clone(),
+                process: child_process_id,
+            })
+            .body(message.body())
+            .send_and_await_response(test_timeout)??;
+
+        if response.is_request() {
+            fail!("tester");
+        };
+        let TesterResponse::Run(result) = response.body().try_into()?;
+        if let Err(FailResponse { test, file, line, column }) = result {
+            fail!(test, file, line, column);
+        }
+    }
+
+    println!("test_runner: done running {:?}", children);
+
+    Response::new()
+        .body(TesterResponse::Run(Ok(())))
+        .send()?;
+
+    Ok(())
+}
+
 fn handle_message(
     our: &Address,
-    _messages: &mut Messages,
     node_names: &mut Vec<String>,
 ) -> anyhow::Result<()> {
     let Ok(message) = await_message() else {
         return Ok(());
     };
 
-    match message {
-        Message::Response { source, body, .. } => {
-            match serde_json::from_slice(&body)? {
-                tt::TesterResponse::Pass | tt::TesterResponse::Fail { .. } => {
-                    if (source.process.package_name != "tester")
-                        | (source.process.publisher_node != "sys")
-                    {
-                        return Err(tt::TesterError::UnexpectedResponse.into());
-                    }
-                    Response::new().body(body).send().unwrap();
-                }
-                tt::TesterResponse::GetFullMessage(_) => {
-                    fail!("tester");
-                }
-            }
-            Ok(())
-        }
-        Message::Request { body, .. } => {
-            match serde_json::from_slice(&body)? {
-                tt::TesterRequest::Run {
-                    input_node_names,
-                    test_timeout,
-                    ..
-                } => {
-                    println!("got Run");
-
-                    assert!(input_node_names.len() >= 1);
-                    *node_names = input_node_names.clone();
-
-                    if our.node != node_names[0] {
-                        Response::new()
-                            .body(serde_json::to_vec(&tt::TesterResponse::Pass).unwrap())
-                            .send()
-                            .unwrap();
-                    } else {
-                        // we are master node
-                        let child = "/tester:sys/pkg/test_runner.wasm";
-                        let child_process_id = match spawn(
-                            None,
-                            child,
-                            OnExit::None, //  TODO: notify us
-                            our_capabilities(),
-                            vec!["vfs:distro:sys".parse::<ProcessId>().unwrap()],
-                            false, // not public
-                        ) {
-                            Ok(child_process_id) => child_process_id,
-                            Err(e) => {
-                                println!("couldn't spawn {}: {}", child, e);
-                                fail!("tester");
-                            }
-                        };
-                        Request::new()
-                            .target(Address {
-                                node: our.node.clone(),
-                                process: child_process_id,
-                            })
-                            .body(body)
-                            .expects_response(test_timeout * 2)
-                            .send()?;
-                    }
-                }
-                tt::TesterRequest::KernelMessage(_) | tt::TesterRequest::GetFullMessage(_) => {
-                    fail!("tester");
-                }
-            }
-            Ok(())
-        }
+    if !message.is_request() {
+        return handle_response(&message);
     }
+    return handle_request(our, &message, node_names);
 }
 
 call_init!(init);
 fn init(our: Address) {
-    let mut messages: Messages = IndexMap::new();
     let mut node_names: Vec<String> = Vec::new();
     match Request::new()
         .target(make_vfs_address(&our).unwrap())
@@ -154,7 +243,7 @@ fn init(our: Address) {
     }
 
     loop {
-        match handle_message(&our, &mut messages, &mut node_names) {
+        match handle_message(&our, &mut node_names) {
             Ok(()) => {}
             Err(e) => {
                 println!("tester: error: {:?}", e,);
