+++ kinode/packages/tester/tester/src/tester_lib.rs
@@ -1,0 +1,1 @@
+../../tester_lib.rs
