--- kinode/packages/tester/tester/src/tester_types.rs
@@ -1,1 +1,0 @@
-../../tester_types.rs
